Refactor
- Move Orbital, Surface, OabStage and MicroEntries to their respective Entries\<file>.cs
- Remove MetersTODistanceString (MicroUtility) to clean up more
